fix: 회원가입 types 추가
