+++ types/SignUpResponse.ts
@@ -1,0 +1,2 @@
+export interface SignUpResponse {
+}
